requirements.txt
selenium==3.141.0
requests==2.22.0
slackclient==2.3.0